keystone: Fixes for move to mariadb operator
We were creating the CR from the default in the repo. This default
still assumed the previous default mariadb service name.

We pull the CR into the ansible repo and correct the db service name.
This makes more sense than trying to ensure example data is consistent
across multiple repos.

We also ensure the playbook waits for the deployment to complete
before continuing, which was causing errors deploying subsequent
control plane components.