Makefile: Create untracked files for common local-only changes
This commit addresses the problem of making local changes to
vars/defaults.yaml and hosts, both of which are currently tracked by
git. We never want local changes to these files to be committed
accidentally, and we don't want conflicts caused to local changes.

We create local-defaults.yaml for overriding variables in
vars/defaults.yaml. If it doesn't exist already, make will initialise
an empty file. local-defaults.yaml has the same format as
vars/defaults.yaml. You can copy individual variables there and change
them.

We remove hosts from git. Make will initialise it with the current
contents (defaulting to running against localhost) if it does not
exist, but it will not be subsequently updated if it is changed.