compute node: convert to oc_local and make idempotent
The playbook was not previously idempotent, as it would decrease the
size of the worker-0 by osp_compute_scale on every execution. We make
it idempotent by:

* Moving the CR into dev-tools, and setting the desired number of
  workers directly
* Unconditionally scaling worker-0 to the total number of remaining
  baremetalhosts

Also, as the operator is now deployed by OLM there is no need to wait
for it in this playbook: the OLM playbook does that for us.

We also create the openstack client config required by the compute
node operator. This needs to move elsewhere, but can live here for
now. If it's generally required, it should probably be created by the
keystone operator.

Note that the compute does not yet become Ready on my cluster, and
therefore the wait at the end of this playbook will always fail. That
is likely unrelated to this change.