nfs: Simplify and fix connection=local issues
Similar to oc_local, something in my environment means that the
behaviour of tasks using connection=local had changed, and this no
longer worked.

The external task was intended to be reusable, but that doesn't seem
important at the moment. Pull it into the playbook and split it into
separate plays running on convergence_base and localhost.