nova: Make compute service names prettier
Pods created by a Deployment have a hash in their name. Pods created
by a StatefulSet have a consistent name with sequential numbering.
This change switches to consistent naming for pods whose hostname is
presented via the openstack api. As well as looking nicer, it more
importantly means we don't need to cleanup the compute service entry
in OpenStack every time we restart a pod, only when we scale the
service down.