prepare_host: Remove unnecessary host configuration
This is a cleanup of openstack and workstation configuration on the
OCP host which are no longer required. This may not quite be a minimal
set, as I believe dev-scripts will install podman and buildah.
However, this is the set I tested.

This also removes some workstation tools we were installing on the OCP
host under the assumption that it is also used as a workstation. Any
required tools should be added back to 04_local_oc_client.yaml,
although I don't believe there are any. Personal preference tools can
be installed manually on individual workstations.